requirements.txt
acme==1.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aioredis==1.3.1
aiosignal==1.4.0
alembic==1.11.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.8.1
asn1crypto==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
Babel==2.8.0
backports.tarfile==1.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.9.0
Brlapi==0.8.3
build==1.2.2.post1
CacheControl==0.14.3
cached-property==1.5.2
cachetools==5.5.2
cbor2==5.4.2
certbot==1.21.0
certbot-apache==1.21.0
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
cmdstanpy==1.2.5
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
contourpy==1.3.1
crashtest==0.4.1
cryptography==3.4.8
cupshelpers==1.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databases==0.9.0
dbus-python==1.2.18
debugpy==1.8.12
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.6
djangorestframework==3.15.2
dnspython==2.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
dulwich==0.22.8
duplicity==0.8.21
ecdsa==0.19.1
email_validator==2.2.0
et-xmlfile==1.0.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.103.0
fastapi-cache==0.1.0
fasteners==0.14.1
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==3.0.3
fonttools==4.55.8
freezegun==1.0.0
frozenlist==1.7.0
future==0.18.2
gbinder-python==1.1.2
geoip2==2.9.0
gevent==21.8.0
git-filter-repo==2.47.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-speech==2.33.0
googleapis-common-protos==1.70.0
googletrans==3.1.0a0
greenlet==1.1.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.9.0
h2==3.2.0
h5py==3.12.1
hiredis==3.2.1
holidays==0.75
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==1.1.1
installer==0.7.0
ipykernel==6.29.5
ipython==8.32.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jdcal==1.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.4.2
josepy==1.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libsass==0.20.1
libvirt-python==8.0.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
maxminddb==2.0.3
monotonic==1.6
more-itertools==8.10.0
msgpack==1.1.0
multidict==6.6.3
narwhals==1.27.1
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
nltk==3.9.1
num2words==0.5.10
numpy==1.23.5
oauthlib==3.2.0
odoo==18.0.post20250707
ofxparse==0.21
olefile==0.46
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.0.9
packaging==24.2
palmerpenguins==0.1.4
pandas==1.5.3
paramiko==2.9.3
parsedatetime==2.6
parso==0.8.4
passlib==1.7.4
pbs-installer==2025.5.17
pexpect==4.8.0
Pillow==9.0.1
pkginfo==1.12.1.2
platformdirs==4.3.6
plotly==5.15.0
pluggy==0.13.0
poetry==2.1.3
poetry-core==2.1.3
polib==1.1.1
prompt_toolkit==3.0.50
propcache==0.3.2
prophet==1.1.5
proto-plus==1.26.1
protobuf==6.31.1
psutil==6.1.1
psycopg2==2.9.2
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.10.0
pyasn1==0.6.1
pyasn1-modules==0.2.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.19.1
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyproject_hooks==1.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
pytest==6.2.5
python-apt==2.4.0+ubuntu4
python-augeas==0.5.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-ldap==3.2.0
python-multipart==0.0.20
python-stdnum==1.17
pytz==2022.1
pyusb==1.2.1.post1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.1
qrcode==7.3.1
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
retrying==1.4.0
rfc3986==1.5.0
rjsmin==1.1.0
roman==3.3
rpds-py==0.22.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
ssh-import-id==5.11
stack-data==0.6.3
stanio==0.5.1
starlette==0.27.0
systemd-python==234
tenacity==9.1.2
texttable==1.6.4
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.5.9.12
twilio==8.11.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.23.2
virtualenv==20.31.2
vobject==0.9.6.1
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==3.0.6
xdg==5
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
yarl==1.20.1
zeep==4.1.0
zipp==3.22.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
zstandard==0.23.0
